+++ Common/IntCodes/IntCodeComputer.cs
@@ -8,20 +8,20 @@ namespace Common.IntCodes
 {
     public class IntCodeComputer
     {
-        public IntCodeState Parse(string inputProgram, Func<int> getNextInputValue, Action<int>? onNewOutputValue) => new IntCodeState(
+        public IntCodeState Parse(string inputProgram, Func<long> getNextInputValue, Action<long>? onNewOutputValue) => new IntCodeState(
             inputProgram.Split(',')
-                .Select(int.Parse)
+                .Select(long.Parse)
                 .ToArray(),
             getNextInputValue,
             onNewOutputValue);
 
-        public IntCodeState ParseAndEvaluate(string inputProgram, params int[]? inputValues)
+        public IntCodeState ParseAndEvaluate(string inputProgram, params long[]? inputValues)
         {
-            var inputValuesQueue = new Queue<int>(inputValues ?? Array.Empty<int>());
+            var inputValuesQueue = new Queue<long>(inputValues ?? Array.Empty<long>());
             return ParseAndEvaluateWithSignalling(inputProgram, () => inputValuesQueue.Dequeue(), null);
         }
 
-        public IntCodeState ParseAndEvaluateWithSignalling(string inputProgram, Func<int> receiveInputValue, Action<int>? sendOutputValue)
+        public IntCodeState ParseAndEvaluateWithSignalling(string inputProgram, Func<long> receiveInputValue, Action<long>? sendOutputValue)
         {
             var intCodeState = Parse(inputProgram, receiveInputValue, sendOutputValue);
 
@@ -36,7 +36,7 @@ namespace Common.IntCodes
             return intCodeState;
         }
 
-        private static int? EvalInstruction(Instruction instruction) =>
+        private static long? EvalInstruction(Instruction instruction) =>
             instruction.OpCode switch
                 {
                 1 => EvalMathInstruction(instruction),
@@ -50,7 +50,7 @@ namespace Common.IntCodes
                 _ => throw new InvalidOperationException("Invalid opCode: " + instruction.OpCode)
                 };
 
-        private static int? EvalMathInstruction(Instruction instruction)
+        private static long? EvalMathInstruction(Instruction instruction)
         {
             var param1 = instruction.GetParamUsingMode(0);
             var param2 = instruction.GetParamUsingMode(1);
@@ -69,7 +69,7 @@ namespace Common.IntCodes
             return null;
         }
 
-        private static int? EvalInputInstruction(Instruction instruction)
+        private static long? EvalInputInstruction(Instruction instruction)
         {
             // opCode 3 takes a single integer as input and saves it to the address given by its only parameter.
             // For example, the instruction 3,50 would take an input value and store it at address 50.
@@ -79,7 +79,7 @@ namespace Common.IntCodes
             return null;
         }
 
-        private static int? EvalOutputInstruction(Instruction instruction)
+        private static long? EvalOutputInstruction(Instruction instruction)
         {
             // opCode 4 outputs the value of its only parameter.
             // For example, the instruction 4,50 would output the value at address 50.
@@ -90,7 +90,7 @@ namespace Common.IntCodes
             return null;
         }
 
-        private static int? EvalJumpInstruction(Instruction instruction)
+        private static long? EvalJumpInstruction(Instruction instruction)
         {
             // Opcode 5 is jump-if-true: if the first parameter is non-zero, it sets the instruction pointer to
             // the value from the second parameter. Otherwise, it does nothing.
@@ -112,7 +112,7 @@ namespace Common.IntCodes
             }
         }
 
-        private static int? EvalRelationalInstruction(Instruction instruction)
+        private static long? EvalRelationalInstruction(Instruction instruction)
         {
             // Opcode 7 is less than: if the first parameter is less than the second parameter,
             // it stores 1 in the position given by the third parameter. Otherwise, it stores 0.
@@ -139,14 +139,14 @@ namespace Common.IntCodes
 
         #region Signalling, PhaseSettingSequences and FeedbackLoops
 
-        public int ParseAndEvaluateWithPhaseSettingSequenceAndFeedbackLoop(string inputProgram, int[] phaseSettingSequence)
+        public long ParseAndEvaluateWithPhaseSettingSequenceAndFeedbackLoop(string inputProgram, int[] phaseSettingSequence)
         {
             var deviceSignalConnectors = phaseSettingSequence.Select(phaseSetting => new PhaseSignalConnector(phaseSetting))
                 .ToArray();
 
             deviceSignalConnectors.First().SetNextValue(0); // Seed input signal of the first device
 
-            var finalResults = new int[phaseSettingSequence.Length];
+            var finalResults = new long[phaseSettingSequence.Length];
 
             Parallel.ForEach(
                 deviceSignalConnectors.Select((connector, index) => (connector, index)),
@@ -172,15 +172,15 @@ namespace Common.IntCodes
         {
             private readonly AutoResetEvent waitHandle = new AutoResetEvent(false);
 
-            private int nextValue;
+            private long nextValue;
 
-            public virtual int ReceiveNextValue()
+            public virtual long ReceiveNextValue()
             {
                 waitHandle.WaitOne();
                 return nextValue;
             }
 
-            public void SetNextValue(int value)
+            public void SetNextValue(long value)
             {
                 nextValue = value;
                 waitHandle.Set();
@@ -197,7 +197,7 @@ namespace Common.IntCodes
                 this.phaseSetting = phaseSetting;
             }
 
-            public override int ReceiveNextValue()
+            public override long ReceiveNextValue()
             {
                 if (phaseSettingUsed)
                 {
