+++ Day6/Day6Solver.cs
@@ -6,7 +6,7 @@ namespace Day6
 {
     public class Day6Solver : SolverReadAllLines
     {
-        public override int? SolvePart1(string[] mapLines)
+        public override long? SolvePart1(string[] mapLines)
         {
             // What is the total number of direct and indirect orbits?
             var nodeDictionary = BuildHierarchy(mapLines);
@@ -41,7 +41,7 @@ namespace Day6
             return node;
         }
 
-        public override int? SolvePart2(string[] mapLines)
+        public override long? SolvePart2(string[] mapLines)
         {
             // What is the minimum number of orbital transfers required to move from the object YOU are orbiting to the object SAN is orbiting?
             // (Between the objects they are orbiting - not between YOU and SAN.)
