+++ Day9/Day9Solver.cs
@@ -4,12 +4,12 @@ namespace Day9
 {
     public class Day9Solver : SolverReadAllLines
     {
-        public override int? SolvePart1(string[] input)
+        public override long? SolvePart1(string[] input)
         {
             return base.SolvePart1(input);
         }
 
-        public override int? SolvePart2(string[] input)
+        public override long? SolvePart2(string[] input)
         {
             return base.SolvePart2(input);
         }
